+++ bzt/jmx/tools.py
@@ -15,7 +15,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 """
-import json
 import os
 import traceback
 from distutils.version import LooseVersion
@@ -23,9 +22,10 @@ from distutils.version import LooseVersion
 from bzt import TaurusInternalException, TaurusConfigError
 from bzt.engine import Scenario
 from bzt.jmx import JMX
+from bzt.jmx.threadgroups import ThreadGroup, ConcurrencyThreadGroup, ThreadGroupHandler
 from bzt.requests_model import RequestVisitor, has_variable_pattern
 from bzt.six import etree, iteritems, numeric_types
-from bzt.utils import BetterDict, dehumanize_time, ensure_is_dict, get_host_ips, get_full_path, guess_csv_dialect
+from bzt.utils import BetterDict, dehumanize_time, ensure_is_dict, get_full_path, guess_csv_dialect, load_class
 
 
 class RequestCompiler(RequestVisitor):
@@ -34,7 +34,7 @@ class RequestCompiler(RequestVisitor):
         self.jmx_builder = jmx_builder
 
     def visit_hierarchichttprequest(self, request):
-        return self.jmx_builder.compile_http_request(request)
+        return self.jmx_builder.compile_request(request)
 
     def visit_ifblock(self, block):
         return self.jmx_builder.compile_if_block(block)
@@ -64,222 +64,6 @@ class RequestCompiler(RequestVisitor):
 
     def visit_setvariables(self, block):
         return self.jmx_builder.compile_set_variables_block(block)
-
-
-class AbstractThreadGroup(object):
-    XPATH = None
-    RAMP_UP_SEL = None
-    CONCURRENCY_SEL = None
-
-    def __init__(self, element, logger):
-        self.element = element
-        self.gtype = self.__class__.__name__
-        self.log = logger.getChild(self.gtype)
-
-    def get_testname(self):
-        return self.element.get('testname')
-
-    def set_concurrency(self, concurrency=None):
-        self.log.warning('Setting of concurrency for %s not implemented', self.gtype)
-
-    def set_ramp_up(self, ramp_up=None):
-        self.log.warning('Setting of ramp-up for %s not implemented', self.gtype)
-
-    def get_duration(self):
-        """
-        task duration or None if getting isn't possible (skipped, timeless, jmeter variables, etc.)
-        """
-        self.log.warning('Getting of duration for %s not implemented', self.gtype)
-
-    def get_rate(self, pure=False):
-        self.log.warning('Getting of rate for %s not implemented', self.gtype)
-
-    def get_iterations(self):
-        """
-        iterations number or None if getting isn't possible (skipped, unsupported, jmeter variables, etc.)
-        Note: ConcurrencyThreadGroup and ArrivalsThreadGroup aren't stopped by iterations limit
-        """
-        self.log.warning('Getting of iterations for %s not implemented', self.gtype)
-
-    def get_ramp_up(self, pure=False):
-        if not self.RAMP_UP_SEL:
-            self.log.warning('Getting of ramp-up for %s not implemented', self.gtype)
-            return 1
-
-        return self._get_val(self.RAMP_UP_SEL, name='ramp-up', default=0, pure=pure)
-
-    def get_concurrency(self, pure=False):
-        if not self.CONCURRENCY_SEL:
-            self.log.warning('Getting of concurrency for %s not implemented', self.gtype)
-            return 1
-
-        return self._get_val(self.CONCURRENCY_SEL, name='concurrency', default=1, pure=pure)
-
-    def _get_val(self, selector, name='', default=None, convertor=int, pure=False):
-        element = self.element.find(selector)
-        if element is None:
-            string_val = None
-        else:
-            string_val = element.text
-
-        if pure:
-            return string_val
-
-        try:
-            return convertor(string_val)
-        except (ValueError, TypeError):
-            if default:
-                msg = "Parsing {param} '{val}' in group '{gtype}' failed, choose {default}"
-                self.log.warning(msg.format(param=name, val=string_val, gtype=self.gtype, default=default))
-                return default
-
-    def get_on_error(self):
-        action = self.element.find(".//stringProp[@name='ThreadGroup.on_sample_error']")
-        if action is not None:
-            return action.text
-
-
-class ThreadGroup(AbstractThreadGroup):
-    XPATH = 'jmeterTestPlan>hashTree>hashTree>ThreadGroup'
-    CONCURRENCY_SEL = ".//*[@name='ThreadGroup.num_threads']"
-    RAMP_UP_SEL = ".//*[@name='ThreadGroup.ramp_time']"
-
-    def get_duration(self):
-        sched_sel = ".//*[@name='ThreadGroup.scheduler']"
-        scheduler = self._get_val(sched_sel, "scheduler", pure=True)
-
-        if scheduler == 'true':
-            duration_sel = ".//*[@name='ThreadGroup.duration']"
-            return self._get_val(duration_sel, "duration")
-        elif scheduler == 'false':
-            return self._get_val(self.RAMP_UP_SEL, "ramp-up")
-        else:
-            msg = 'Getting of ramp-up for %s is impossible due to scheduler: %s'
-            self.log.warning(msg, (self.gtype, scheduler))
-
-    def get_iterations(self):
-        loop_control_sel = ".//*[@name='LoopController.continue_forever']"
-        loop_controller = self._get_val(loop_control_sel, name="loop controller", pure=True)
-        if loop_controller == "false":
-            loop_sel = ".//*[@name='LoopController.loops']"
-            return self._get_val(loop_sel, name="loops")
-        else:
-            msg = 'Getting of ramp-up for %s is impossible due to loop_controller: %s'
-            self.log.warning(msg, (self.gtype, loop_controller))
-
-
-class SteppingThreadGroup(AbstractThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>kg\.apc\.jmeter\.threads\.SteppingThreadGroup'
-    CONCURRENCY_SEL = ".//*[@name='ThreadGroup.num_threads']"
-
-
-class UltimateThreadGroup(AbstractThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>kg\.apc\.jmeter\.threads\.UltimateThreadGroup'
-
-
-# parent of ConcurrencyThreadGroup and ArrivalThreadGroup
-class AbstractDynamicThreadGroup(AbstractThreadGroup):
-    RAMP_UP_SEL = ".//*[@name='RampUp']"
-
-    def _get_time_unit(self):
-        unit_sel = ".//*[@name='Unit']"
-        return self._get_val(unit_sel, name="unit", pure=True)
-
-    def set_ramp_up(self, ramp_up=None):
-        ramp_up_element = self.element.find(self.RAMP_UP_SEL)
-        ramp_up_element.text = str(ramp_up)
-
-    def get_duration(self):
-        hold_sel = ".//*[@name='Hold']"
-
-        hold = self._get_val(hold_sel, name="hold")
-        ramp_up = self.get_ramp_up()
-
-        # 'empty' means 0 sec, let's detect that
-        p_hold = self._get_val(hold_sel, name="hold", pure=True)
-        p_ramp_up = self.get_ramp_up(pure=True)
-        if hold is None and not p_hold:
-            hold = 0
-        if ramp_up is None and not p_ramp_up:
-            ramp_up = 0
-
-        if hold is not None and ramp_up is not None:
-            result = hold + ramp_up
-            if self._get_time_unit() == 'M':
-                result *= 60
-
-            return result
-
-    def get_iterations(self):
-        iter_sel = ".//*[@name='Iterations']"
-        return self._get_val(iter_sel, name="iterations")
-
-
-class ConcurrencyThreadGroup(AbstractDynamicThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>com\.blazemeter\.jmeter\.threads\.concurrency\.ConcurrencyThreadGroup'
-    CONCURRENCY_SEL = ".//*[@name='TargetLevel']"
-
-    def set_concurrency(self, concurrency=None):
-        concurrency_prop = self.element.find(self.CONCURRENCY_SEL)
-        concurrency_prop.text = str(concurrency)
-
-
-class ArrivalsThreadGroup(AbstractDynamicThreadGroup):
-    XPATH = r'jmeterTestPlan>hashTree>hashTree>com\.blazemeter\.jmeter\.threads\.arrivals\.ArrivalsThreadGroup'
-    RATE_SEL = ".//*[@name='TargetLevel']"
-
-    def get_rate(self, pure=False):
-        return self._get_val(self.RATE_SEL, name='rate', default=1, pure=pure)
-
-    def set_rate(self, rate=None):
-        rate_prop = self.element.find(self.RATE_SEL)
-        rate_prop.text = str(rate)
-
-
-class ThreadGroupHandler(object):
-    CLASSES = [ThreadGroup, SteppingThreadGroup, UltimateThreadGroup, ConcurrencyThreadGroup, ArrivalsThreadGroup]
-
-    def __init__(self, logger):
-        self.log = logger.getChild(self.__class__.__name__)
-
-    def groups(self, jmx):
-        """
-        Get wrappers for thread groups that are enabled
-        """
-        for _class in self.CLASSES:
-            for group in jmx.get(_class.XPATH):
-                if group.get("enabled") != "false":
-                    yield _class(group, self.log)
-
-    def convert(self, group, target, load, concurrency):
-        """
-        Convert a thread group to ThreadGroup/ConcurrencyThreadGroup for applying of load
-        """
-        msg = "Converting %s (%s) to %s and apply load parameters"
-        self.log.debug(msg, group.gtype, group.get_testname(), target)
-        on_error = group.get_on_error()
-
-        if target == ThreadGroup.__name__:
-            new_group_element = JMX.get_thread_group(
-                concurrency=concurrency,
-                rampup=load.ramp_up,
-                hold=load.hold,
-                iterations=load.iterations,
-                testname=group.get_testname(),
-                on_error=on_error)
-        elif target == ConcurrencyThreadGroup.__name__:
-            new_group_element = JMX.get_concurrency_thread_group(
-                concurrency=concurrency,
-                rampup=load.ramp_up,
-                hold=load.hold,
-                steps=load.steps,
-                testname=group.get_testname(),
-                on_error=on_error)
-        else:
-            self.log.warning('Unsupported preferred thread group: %s', target)
-            return
-
-        group.element.getparent().replace(group.element, new_group_element)
 
 
 class LoadSettingsProcessor(object):
@@ -391,38 +175,20 @@ class LoadSettingsProcessor(object):
         jmx.append(JMeterScenarioBuilder.TEST_PLAN_SEL, etree.Element("hashTree"))
 
 
-class JMeterScenarioBuilder(JMX):
-    """
-    Helper to build JMeter test plan from Scenario
+class ProtocolHandler(object):
 
-    :param executor: ScenarioExecutor
-    :param original: inherited from JMX
-    """
+    def __init__(self, sys_props):
+        super(ProtocolHandler, self).__init__()
+        self.system_props = sys_props
 
-    def __init__(self, executor, original=None):
-        super(JMeterScenarioBuilder, self).__init__(original)
-        self.executor = executor
-        self.scenario = executor.get_scenario()
-        self.engine = executor.engine
-        self.system_props = BetterDict()
-        self.request_compiler = None
+    def get_toplevel_elements(self, scenario):
+        return []
 
-    def __gen_managers(self, scenario):
-        elements = []
-        if scenario.get("store-cache", True):
-            elements.append(self._get_cache_mgr())
-            elements.append(etree.Element("hashTree"))
-        if scenario.get("store-cookie", True):
-            elements.append(self._get_cookie_mgr(scenario))
-            elements.append(etree.Element("hashTree"))
-        if scenario.get("use-dns-cache-mgr", True):
-            elements.append(self.get_dns_cache_mgr())
-            elements.append(etree.Element("hashTree"))
-            self.system_props.merge({"system-properties": {"sun.net.inetaddr.ttl": 0}})
-        return elements
+    def get_sampler_pair(self, scenario, request):
+        return None, None
 
     @staticmethod
-    def smart_time(any_time):  # FIXME: bad name for the function, does not reflect what it does
+    def safe_time(any_time):
         try:
             smart_time = int(1000 * dehumanize_time(any_time))
         except TaurusInternalException:
@@ -430,25 +196,35 @@ class JMeterScenarioBuilder(JMX):
 
         return smart_time
 
-    def __gen_defaults(self, scenario):
-        default_address = scenario.get("default-address", None)
-        retrieve_resources = scenario.get("retrieve-resources", True)
-        resources_regex = scenario.get("retrieve-resources-regex", None)
-        concurrent_pool_size = scenario.get("concurrent-pool-size", 4)
 
-        content_encoding = scenario.get("content-encoding", None)
+class JMeterScenarioBuilder(JMX):
+    """
+    Helper to build JMeter test plan from Scenario
 
-        timeout = scenario.get("timeout", None)
-        timeout = self.smart_time(timeout)
-        elements = [self._get_http_defaults(default_address, timeout, retrieve_resources,
-                                            concurrent_pool_size, content_encoding, resources_regex),
-                    etree.Element("hashTree")]
-        return elements
+    :type protocol_handlers: list[ProtocolHandler]
+    """
+
+    def __init__(self, executor, original=None):
+        """
+        :type executor: ScenarioExecutor
+        :type original: JMX
+        """
+        super(JMeterScenarioBuilder, self).__init__(original)
+        self.executor = executor
+        self.scenario = executor.get_scenario()
+        self.engine = executor.engine
+        self.system_props = BetterDict()
+        self.request_compiler = None
+        self.protocol_handlers = []
+        for cls_name in self.executor.settings.get("protocol-handlers", []):
+            cls_obj = load_class(cls_name)
+            instance = cls_obj(self.system_props)
+            self.protocol_handlers.append(instance)
 
     def __add_think_time(self, children, req):
         think_time = req.priority_option('think-time')
         if think_time is not None:
-            children.append(JMX._get_constant_timer(self.smart_time(think_time)))
+            children.append(JMX._get_constant_timer(ProtocolHandler.safe_time(think_time)))
             children.append(etree.Element("hashTree"))
 
     def __add_extractors(self, children, req):
@@ -588,20 +364,6 @@ class JMeterScenarioBuilder(JMX):
             children.append(JMX._get_jsr223_element(lang, script_file, parameters, execute, script_text))
             children.append(etree.Element("hashTree"))
 
-    def _get_merged_ci_headers(self, req, header):
-        def dic_lower(dic):
-            return {str(k).lower(): str(dic[k]).lower() for k in dic}
-
-        ci_scenario_headers = dic_lower(self.scenario.get_headers())
-        ci_request_headers = dic_lower(req.headers)
-        headers = BetterDict()
-        headers.merge(ci_scenario_headers)
-        headers.merge(ci_request_headers)
-        if header.lower() in headers:
-            return headers[header]
-        else:
-            return None
-
     def __gen_requests(self, scenario):
         requests = scenario.get_requests()
         elements = []
@@ -611,46 +373,33 @@ class JMeterScenarioBuilder(JMX):
 
     def compile_scenario(self, scenario):
         elements = []
-        elements.extend(self.__gen_managers(scenario))
-        elements.extend(self.__gen_defaults(scenario))
+        for protocol in self.protocol_handlers:
+            elements.extend(protocol.get_toplevel_elements(scenario))
         elements.extend(self.__gen_datasources(scenario))
         elements.extend(self.__gen_requests(scenario))
         return elements
 
-    def compile_http_request(self, request):
+    def compile_request(self, request):
         """
 
         :type request: HierarchicHTTPRequest
         :return:
         """
-        timeout = request.priority_option('timeout')
-        if timeout is not None:
-            timeout = self.smart_time(timeout)
+        sampler = children = None
+        for protocol in self.protocol_handlers:
+            sampler, children = protocol.get_sampler_pair(self.scenario, request)
+            if sampler is not None:
+                break
 
-        content_type = self._get_merged_ci_headers(request, 'content-type')
-        if content_type == 'application/json' and isinstance(request.body, (dict, list)):
-            body = json.dumps(request.body)
-        else:
-            body = request.body
-
-        use_random_host_ip = request.priority_option('random-source-ip', default=False)
-        host_ips = get_host_ips(filter_loopbacks=True) if use_random_host_ip else []
-        http = JMX._get_http_request(request.url, request.label, request.method, timeout, body,
-                                     request.priority_option('keepalive', default=True),
-                                     request.upload_files, request.content_encoding,
-                                     request.priority_option('follow-redirects', default=True),
-                                     use_random_host_ip, host_ips)
-
-        children = etree.Element("hashTree")
-
-        if request.headers:
-            children.append(JMX._get_header_mgr(request.headers))
-            children.append(etree.Element("hashTree"))
+        if sampler is None:
+            self.log.warning("Problematic request: %s", request)
+            raise TaurusInternalException("Unable to handle request, please review missing options")
 
         self.__add_think_time(children, request)
 
         self.__add_assertions(children, request)
 
+        timeout = ProtocolHandler.safe_time(request.priority_option('timeout'))
         if timeout is not None:
             children.append(JMX._get_dur_assertion(timeout))
             children.append(etree.Element("hashTree"))
@@ -659,7 +408,7 @@ class JMeterScenarioBuilder(JMX):
 
         self.__add_jsr_elements(children, request)
 
-        return [http, children]
+        return [sampler, children]
 
     def compile_if_block(self, block):
         elements = []
